Change default MySQL client to MariaDB
This PR is a response to pretty catastrophic issue caused by expiring
key on MySQL repository on 14th of December. Oracle does not follow the
best practices for signing their packages (while all others do) and
their packages and repositories are signed with a key with short
expiry date. This basically puts an expiry on their repository, and
anyone who releases images following best practices of installation,
while keeping the repository after installing mysql libraries has to
rebuild their past released images every 2 years or so.

This is the last straw for our MySQL client installation problems (we
had a few, especially for ARM images) and we decided that we will
switch to MariaDB client libraries by default (while allowing our
users to build custom images with MySQL libraries).

The issue tracked in Airflow repository is: #36231

The issues in Oracle's MySQL repo:

* https://bugs.mysql.com/bug.php?id=113427
* https://bugs.mysql.com/bug.php?id=113428
* https://bugs.mysql.com/bug.php?id=113432

This PR implements a number of changes:

* MariaDB client is now default client used for both ARM and X86
* both pre-2023 and 2023 keys for MySQL are now added to be trusted
  when custom image with MySQL client is built
* MySQL repository is removed after installing MySQL (to avoid
  repeating similar fiasco for MySQL users in 2025
* changelog added and instructions on how to build custom image
  with MySQL client
* one of our test suites is converted to use "current" image not
  latest released image (that was bug in our CI).
* test was added in `canary` and `release` builds in CI to also test
  build of custom image with MySQL client